Refactor block variations
